backend/requirements.txt
django
djangorestframework
python-dotenv
openai
django-cors-headers
gunicorn